requirements.txt
Django==1.4
PIL==1.1.7
argparse==1.2.1
django-filebrowser==3.4.2
django-grappelli==2.3.8
docutils==0.9
feedparser==5.1.2
wsgiref==0.1.2
